Created Utilities project + minor refactoring
