+++ Chirp.sln
@@ -14,6 +14,12 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Chirp.CSVDB.Tests", "test\C
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Chirp.DBService", "src\Chirp.DBService\Chirp.DBService.csproj", "{CCB1413A-C7EE-4524-97D1-4F7A5632A792}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Chirp.Utilities.Tests", "test\Chirp.Utilities.Tests\Chirp.Utilities.Tests.csproj", "{455C4F27-453B-49AF-8FFC-9883A9731D6D}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Chirp.Utilities", "src\Chirp.Utilities\Chirp.Utilities.csproj", "{59E69556-BDE6-4D72-8CEE-94A716B58A6A}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Chirp.DBService.Tests", "test\Chirp.DBService.Tests\Chirp.DBService.Tests.csproj", "{5067CA3F-809F-4EB9-B620-6EE34350C522}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -40,6 +46,18 @@ Global
 		{CCB1413A-C7EE-4524-97D1-4F7A5632A792}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{CCB1413A-C7EE-4524-97D1-4F7A5632A792}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{CCB1413A-C7EE-4524-97D1-4F7A5632A792}.Release|Any CPU.Build.0 = Release|Any CPU
+		{455C4F27-453B-49AF-8FFC-9883A9731D6D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{455C4F27-453B-49AF-8FFC-9883A9731D6D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{455C4F27-453B-49AF-8FFC-9883A9731D6D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{455C4F27-453B-49AF-8FFC-9883A9731D6D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{59E69556-BDE6-4D72-8CEE-94A716B58A6A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{59E69556-BDE6-4D72-8CEE-94A716B58A6A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{59E69556-BDE6-4D72-8CEE-94A716B58A6A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{59E69556-BDE6-4D72-8CEE-94A716B58A6A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5067CA3F-809F-4EB9-B620-6EE34350C522}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5067CA3F-809F-4EB9-B620-6EE34350C522}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5067CA3F-809F-4EB9-B620-6EE34350C522}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5067CA3F-809F-4EB9-B620-6EE34350C522}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{803BB908-B214-4531-B55B-437F443C93B6} = {F5652059-FD91-49A9-9280-F29A964C66D4}
@@ -47,5 +65,8 @@ Global
 		{8EB01794-1A77-434A-8C82-CAC4D32B32E3} = {30A1EFEE-14BC-4970-952F-F05F508AD6AD}
 		{ECD1530D-F2C3-488D-8B65-24EDE1895480} = {30A1EFEE-14BC-4970-952F-F05F508AD6AD}
 		{CCB1413A-C7EE-4524-97D1-4F7A5632A792} = {F5652059-FD91-49A9-9280-F29A964C66D4}
+		{455C4F27-453B-49AF-8FFC-9883A9731D6D} = {30A1EFEE-14BC-4970-952F-F05F508AD6AD}
+		{59E69556-BDE6-4D72-8CEE-94A716B58A6A} = {F5652059-FD91-49A9-9280-F29A964C66D4}
+		{5067CA3F-809F-4EB9-B620-6EE34350C522} = {30A1EFEE-14BC-4970-952F-F05F508AD6AD}
 	EndGlobalSection
 EndGlobal
