+++ src/Chirp.CLI.Client/Chirp.CLI.Client.csproj
@@ -13,5 +13,9 @@
         <PackageReference Include="System.CommandLine" Version="2.0.0-beta4.22272.1" />
         <PackageReference Include="System.CommandLine.NamingConventionBinder" Version="2.0.0-beta4.22272.1" />
     </ItemGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\Chirp.Utilities\Chirp.Utilities.csproj" />
+    </ItemGroup>
     
 </Project>
