+++ src/Chirp.CSVDB/IDatabaseRepository.cs
@@ -1,4 +1,6 @@
-﻿namespace Chirp.CSVDB
+﻿using Chirp.Utilities.Models;
+
+namespace Chirp.CSVDB
 {
     public interface IDatabaseRepository
     {
