+++ src/Chirp.Utilities/Models/Cheep.cs
@@ -1,0 +1,3 @@
+namespace Chirp.Utilities.Models;
+
+public record Cheep(string Author, string Message, long Timestamp);
