+++ test/Chirp.DBService.Tests/UnitTest1.cs
@@ -1,0 +1,9 @@
+namespace Chirp.DBService.Tests;
+
+public class UnitTest1
+{
+    [Fact]
+    public void Test1()
+    {
+    }
+}
